+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Release date ??.??.???? - Release ?.?.?
+
 ### Release date 20.03.2025 - Release 2.1.1
 
 - [[#173]](https://github.com/Marcin-Migdal/m-component-library/issues/173) **[PATCH]** Fix Breadcrumb covered label, implement responsiveness and tooltip on clipped crumb text,