+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### Release date ??.??.???? - Release ?.?.?
+### Release date 25.03.2025 - Release 2.1.2
 
 - [[#186]](https://github.com/Marcin-Migdal/m-component-library/issues/186) **[PATCH]** Fix multiple bug 186, 187, 191, export FormErrorsType, Form register and registerBlur type improvement, changeDropdown input optionMenu z-index
 - [[#188]](https://github.com/Marcin-Migdal/m-component-library/issues/188) **[PATCH]** Fix overlay disabled key press, add onClose callback that is called on "esc" button press.